Reworked LED's drawing procedure
Now it goes row-by-row rather than col-by-col. This improves the visual
look of them, since in my scheme LEDs are connected in columns using one
resistor per column. Because of that the LEDs are getting dimmer if the
whole column is enabled.

Also added a new delay, since the original HAL delay may not be shorter
than 1ms.